conducting performance experiments
NOTE: this commit is not meant for merging, it is just to share some
ideas. Please excuse the messy code style ;)

The combination of the following experiments allows me to bring down
cycle time for a process with 20 steps and a payload of ~1K from >5.5s
to <300ms

* Avoid to flush to filesystem on every commit (Filechannel.force()).
This avoids that we trigger flush on every write. Otherwise we have
the cost of the flush in the processing time of every event, which,
since they  are changed leads to prohibitively bad performance and long
processing times. It also leads to very poor resource utilization.
Remving this lets the filesystem optimize when to actually flush to
disk. Note: this has weaker durability guarantees and discussion is
needed whether this is OK or not. For reference, consult the following
excerpt from the kafka manual where you have the same discussion:
https://kafka.apache.org/documentation/#flush.messages
The discussion is around whether considering data "comitted" once it is
in fs cache on at least 2 nodes is a strong enough guarantee or not. It
could also depend on users workloads.
Also note that in my experiment I just comment it out to better
understand the cost an impact. Even if we consider making this
configurable (like Kafka) and turning it off by default for regular data
processing (like Kafka), there could still be situations where we want
to force it. Also, it could be that we want to not flush on every write
but still regulary (time controlled or every X bytes) to prevent the OS
from accumulating too much backlog and keep processing times consistent.

* Reduce size of readbuffer in SegmentReader: Since events are chained,
for each event we have to go through a full `write -> read` cycle. If
the read buffer is too large, the read time will take longer which is
added as a penaltiy when processing each event. The sum of these
penalties are then noticable in the cycle time. In general this is a
tradeoff between amortizing the constant cost of read opertions with
larger read buffer up until where the cost of actually reading the bytes
dominates. You could find a good size for this readbuffer by conducting
an insolated experiment outside of the workflow engine, starting at
read_buffer_size = log_entry_size and then gradually increase it and
find optimal relationship between throughput and latency.

* Snapshot every minute. In my experiment, I see a gradual, slow
increase in cycle time between snapshots. Creating a snapshot brings it
down again. I don't know why and would ahve to investigate. Something is
accumulating somewhere. At the same time, I could not see an increase in
the max of the cycle time when snapshots are taken which means that they
seem to be cheap. Also taking snapshots reagularly will recuce
reprocessing time on node restart or failover. Howeverm there could be
reasons that I am simply unaware of why taking snapshots every minute or
event more frequently might be a bad idea. Also I have not tested with a
large state.

There are a few classes here to try out:
- Start StandaloneBroker
- Deploy processes with io.zeebe.example.workflow.WorkflowDeployer
- Start JobWorker with io.zeebe.example.job.JobWorkerCreator
- Start workflow instances with
  io.zeebe.example.workflow.NonBlockingWorkflowInstanceCompleter

In the io.zeebe.example.workflow.NonBlockingWorkflowInstanceCompleter
you can control the max concurrency to tradeoff latency vs. throughput.